Refactoring of control code.
Have moved configuration to a separate file and made the setup of the buttons more dynamic based on the config.

The socket setup in the index.js file has also been refactored to cut down on the amount of code due to the nature of the events being very similar.