Added incantation input.
Check was put onto the express server so that the password could be tucked away from the view-source (kids might be sneaky, I don't know!).  Will react depending on if the input is successful or not.

Yes, the incantation is in the source code.  No, I really don't care (especially at 2am!)