Disable input and change volume path.
There were issues with playing the buzzer if someone put in a bad input while the video was playing, so now when a video or audio tack plays the input is disabled.

Changed path to the ticktock volume and added properties for the ambient and video so they can be controlled independently.